+++ redasm/formats/pe/pe_analyzer.h
@@ -12,7 +12,7 @@ class PEAnalyzer: public Analyzer
         typedef std::pair<size_t, std::string> APIInfo;
 
     public:
-        PEAnalyzer(u32 petype, DisassemblerAPI* disassembler, const SignatureFiles &signatures);
+        PEAnalyzer(u64 petype, DisassemblerAPI* disassembler, const SignatureFiles &signatures);
         virtual void analyze();
 
     private:
@@ -24,7 +24,7 @@ class PEAnalyzer: public Analyzer
 
     private:
         std::forward_list<APIInfo> m_wndprocapi;
-        u32 m_petype;
+        u64 m_petype;
 };
 
 }
